requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
alembic
pydantic-settings
python-dotenv
uuid-utils
slowapi
python-jose[cryptography]
httpx
aiosqlite
